requirements.txt
apache-airflow==2.11.0
apache-airflow-providers-postgres==6.2.1
apache-airflow-providers-docker==4.4.1
apache-airflow-providers-ssh==4.1.1
psycopg2-binary==2.9.10